+++ packages/fast-build/CHANGELOG.md
@@ -1,8 +1,34 @@
 # Change Log - @microsoft/fast-build
 
-<!-- This log was last generated on Tue, 07 Apr 2026 00:50:48 GMT and should not be manually modified. -->
+<!-- This log was last generated on Fri, 10 Apr 2026 00:21:15 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.3.2
+
+Fri, 10 Apr 2026 00:21:15 GMT
+
+### Patches
+
+- docs: update fast-build README to match fast-element style and clarify testing purpose (7559015+janechu@users.noreply.github.com)
+
+## 0.3.1
+
+Thu, 09 Apr 2026 00:05:34 GMT
+
+### Patches
+
+- fix: support .length property access on arrays in template expressions (7559015+janechu@users.noreply.github.com)
+- fix: parse JSON array and object literals in HTML attribute values (7559015+janechu@users.noreply.github.com)
+
+## 0.3.0
+
+Wed, 08 Apr 2026 00:19:56 GMT
+
+### Minor changes
+
+- feat: root custom elements receive full entry-level state via render_entry_with_locator (7559015+janechu@users.noreply.github.com)
+- feat: strip state-passing {{binding}} attributes from root custom element opening tags when rendering entry HTML (7559015+janechu@users.noreply.github.com)
 
 ## 0.2.0